More docs and cleanup
