+++ encodings/parquet-variant/src/operations.rs
@@ -531,8 +531,8 @@ mod tests {
         );
 
         let variant_view = vortex_arr.as_opt::<Variant>().unwrap();
-        let child = variant_view.child();
-        let inner_pv = child.as_opt::<ParquetVariant>().unwrap();
+        let storage = variant_view.core_storage();
+        let inner_pv = storage.as_opt::<ParquetVariant>().unwrap();
         let mut ctx = LEGACY_SESSION.create_execution_ctx();
         let roundtripped = inner_pv.to_arrow(&mut ctx)?;
         assert_eq!(roundtripped.inner().null_count(), 2);
