+++ vortex-array/src/arrays/variant/vtable/mod.rs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+mod kernel;
 mod operations;
 mod validity;
 
+use kernel::PARENT_KERNELS;
 use prost::Message;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
@@ -17,40 +19,25 @@ use vortex_utils::aliases::hash_set::HashSet;
 use crate::ArrayRef;
 use crate::ExecutionCtx;
 use crate::ExecutionResult;
-use crate::IntoArray;
 use crate::array::Array;
 use crate::array::ArrayId;
 use crate::array::ArrayView;
 use crate::array::EmptyArrayData;
 use crate::array::VTable;
-use crate::arrays::ChunkedArray;
-use crate::arrays::ConstantArray;
-use crate::arrays::Struct;
-use crate::arrays::scalar_fn::ExactScalarFn;
-use crate::arrays::scalar_fn::ScalarFnFactoryExt;
-use crate::arrays::struct_::StructArrayExt;
 use crate::arrays::variant::CORE_STORAGE_SLOT;
 use crate::arrays::variant::NUM_SLOTS;
 use crate::arrays::variant::SHREDDED_SLOT;
 use crate::arrays::variant::SLOT_NAMES;
-use crate::arrays::variant::VariantArrayExt;
 use crate::arrays::variant::compute::rules::RULES;
 use crate::buffer::BufferHandle;
-use crate::builtins::ArrayBuiltins;
 use crate::dtype::DType;
 use crate::dtype::FieldName;
 use crate::dtype::FieldNames;
 use crate::dtype::Nullability;
 use crate::dtype::StructFields;
-use crate::matcher::Matcher;
 use crate::scalar::Scalar;
 use crate::scalar::ScalarValue;
-use crate::scalar_fn::fns::variant_get::VariantGet;
-use crate::scalar_fn::fns::variant_get::VariantGetOptions;
-use crate::scalar_fn::fns::variant_get::VariantPath;
-use crate::scalar_fn::fns::variant_get::VariantPathElement;
 use crate::serde::ArrayChildren;
-use crate::validity::Validity;
 
 /// A [`Variant`]-encoded Vortex array.
 pub type VariantArray = Array<Variant>;
@@ -98,11 +85,6 @@ impl VTable for Variant {
         vortex_ensure!(
             matches!(dtype, DType::Variant(_)),
             "Expected Variant DType, got {dtype}"
-        );
-        vortex_ensure!(
-            matches!(core_storage.dtype(), DType::Variant(_)),
-            "VariantArray core_storage dtype must be Variant, found {}",
-            core_storage.dtype()
         );
         vortex_ensure!(
             core_storage.dtype() == dtype,
@@ -216,164 +198,8 @@ impl VTable for Variant {
         child_idx: usize,
         ctx: &mut ExecutionCtx,
     ) -> VortexResult<Option<ArrayRef>> {
-        let Some(parent) = ExactScalarFn::<VariantGet>::try_match(parent) else {
-            return Ok(None);
-        };
-        if child_idx != 0 || array.core_storage().is::<Variant>() {
-            return Ok(None);
-        }
-
-        let typed = array
-            .shredded()
-            .map(|shredded| {
-                typed_shredded_path(shredded, parent.options.path().elements(), ctx)?
-                    .map(|typed| typed.mask(array.core_storage().is_not_null()?))
-                    .transpose()
-            })
-            .transpose()?
-            .flatten();
-
-        let Some(typed) = typed else {
-            return execute_fallback_variant_get(
-                array.len(),
-                parent.options.clone(),
-                array.core_storage().clone(),
-                ctx,
-            )
-            .map(Some);
-        };
-        if typed.dtype().is_variant()
-            && parent
-                .options
-                .dtype()
-                .is_some_and(|dtype| !dtype.is_variant())
-        {
-            return execute_fallback_variant_get(
-                array.len(),
-                VariantGetOptions::new(VariantPath::root(), parent.options.dtype().cloned()),
-                typed,
-                ctx,
-            )
-            .map(Some);
-        }
-        if parent.options.dtype().is_none_or(DType::is_variant) {
-            let fallback = match typed.dtype() {
-                DType::Struct(..) => Some(execute_fallback_variant_get(
-                    array.len(),
-                    parent.options.clone(),
-                    array.core_storage().clone(),
-                    ctx,
-                )?),
-                DType::List(..) | DType::FixedSizeList(..) => {
-                    return execute_fallback_variant_get(
-                        array.len(),
-                        parent.options.clone(),
-                        array.core_storage().clone(),
-                        ctx,
-                    )
-                    .map(Some);
-                }
-                _ if all_valid(&typed, ctx)? => None,
-                _ => Some(execute_fallback_variant_get(
-                    array.len(),
-                    parent.options.clone(),
-                    array.core_storage().clone(),
-                    ctx,
-                )?),
-            };
-            return merge_typed_as_variant(typed, fallback, ctx).map(Some);
-        }
-
-        let requested_dtype = parent
-            .options
-            .dtype()
-            .vortex_expect("variant dtype handled above");
-        if typed.dtype().as_nullable() != requested_dtype.as_nullable() {
-            return execute_fallback_variant_get(
-                array.len(),
-                parent.options.clone(),
-                array.core_storage().clone(),
-                ctx,
-            )
-            .map(Some);
-        }
-
-        let typed = typed.cast(parent.dtype().clone())?;
-        if all_valid(&typed, ctx)? {
-            return Ok(Some(typed));
-        }
-
-        let fallback = execute_fallback_variant_get(
-            array.len(),
-            parent.options.clone(),
-            array.core_storage().clone(),
-            ctx,
-        )?;
-        let typed_mask = typed.is_not_null()?;
-        typed_mask
-            .zip(typed, fallback)?
-            .execute::<ArrayRef>(ctx)
-            .map(Some)
+        PARENT_KERNELS.execute(array, parent, child_idx, ctx)
     }
-}
-
-fn typed_shredded_path(
-    shredded: &ArrayRef,
-    path: &[VariantPathElement],
-    ctx: &mut ExecutionCtx,
-) -> VortexResult<Option<ArrayRef>> {
-    let mut current = shredded.clone();
-    for element in path {
-        let VariantPathElement::Field(name) = element else {
-            return Ok(None);
-        };
-        let DType::Struct(..) = current.dtype() else {
-            return Ok(None);
-        };
-        let current_struct = current.execute::<Array<Struct>>(ctx)?;
-        let Some(field) = current_struct.unmasked_field_by_name_opt(name.as_ref()) else {
-            return Ok(None);
-        };
-        current = mask_with_validity(field.clone(), current_struct.validity()?)?;
-    }
-
-    Ok(Some(current))
-}
-
-fn mask_with_validity(array: ArrayRef, validity: Validity) -> VortexResult<ArrayRef> {
-    if validity.no_nulls() {
-        return Ok(array);
-    }
-
-    let len = array.len();
-    array.mask(validity.to_array(len))
-}
-
-fn all_valid(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<bool> {
-    Ok(array.validity()?.execute_mask(array.len(), ctx)?.all_true())
-}
-
-fn merge_typed_as_variant(
-    typed: ArrayRef,
-    fallback: Option<ArrayRef>,
-    ctx: &mut ExecutionCtx,
-) -> VortexResult<ArrayRef> {
-    let dtype = DType::Variant(Nullability::Nullable);
-    let mut chunks = Vec::with_capacity(typed.len());
-
-    for idx in 0..typed.len() {
-        let typed_scalar = typed.execute_scalar(idx, ctx)?;
-        let fallback_scalar = fallback
-            .as_ref()
-            .map(|fallback| fallback.execute_scalar(idx, ctx))
-            .transpose()?;
-        let scalar = merge_typed_scalar_as_variant(typed_scalar, fallback_scalar, &dtype)?;
-
-        chunks.push(ConstantArray::new(scalar, 1).into_array());
-    }
-
-    let core_storage = ChunkedArray::try_new(chunks, dtype)?.into_array();
-    VariantArray::try_new(core_storage, None).map(|array| array.into_array())
 }
 
 fn merge_typed_scalar_as_variant(
@@ -396,7 +222,11 @@ fn merge_typed_scalar_as_variant(
         Scalar::variant(typed_scalar)
     };
 
-    scalar.cast(dtype)
+    if scalar.dtype() == dtype {
+        Ok(scalar)
+    } else {
+        scalar.cast(dtype)
+    }
 }
 
 fn typed_list_as_variant_payload(typed_scalar: Scalar) -> VortexResult<Scalar> {
@@ -504,17 +334,6 @@ fn merge_struct_payload(typed: &Scalar, raw: Option<&Scalar>) -> VortexResult<Sc
         DType::Struct(fields, Nullability::NonNullable),
         Some(ScalarValue::Tuple(values)),
     )
-}
-
-fn execute_fallback_variant_get(
-    len: usize,
-    options: VariantGetOptions,
-    core_storage: ArrayRef,
-    ctx: &mut ExecutionCtx,
-) -> VortexResult<ArrayRef> {
-    VariantGet
-        .try_new_array(len, options, [core_storage])?
-        .execute::<ArrayRef>(ctx)
 }
 
 #[cfg(test)]
